# Release v0.4.3.3b 2025-01-13
## New Features
1. improve conversions (v0.4.3.3b 2025-01-05)
2. add per key space configuration in the `configuration` key space (v0.4.3.3b 2025-01-06)
   currently only `shards` or shard count is configurable
   ```redis
      SELECT configuration
      SET mys.shards 1
      SELECT mys
      SET foo bar
      SELECT configuration
      GET mys.shards
   ```
3. add python/hints example (v0.4.3.3b 2025-01-06)
4. add multi key operations (v0.4.3.3b 2025-01-08)
   `SET "123 abc"` will create a key composed of a number and a string
   `SET "word 12"` can for instance be used to store a text index entry uniquely
5. add `ordered` to per space configuration (v0.4.3.3b 2025-01-10)
6. improve allocation (v0.4.3.3b 2025-01-10)
7. add large geo coding example (v0.4.3.3b 2025-01-12)
8. add standalone RESP server in python example - with messages passed using lbpop (v0.4.3.3b 2025-01-12)
9. rework replication (v0.4.3.3b 2025-01-13)

## Fixes
1. fix conversions (v0.4.3.3b 2025-01-05)
2. fix bloom filter bug when dependency is active (v0.4.3.3b 2025-01-06)
3. fix conversion bugs (v0.4.3.3b 2025-01-09)
4. fix multi key operations (v0.4.3.3b 2025-01-10)
5. fix string trimming (v0.4.3.3b 2025-01-10)
6. fix string decoding (v0.4.3.3b 2025-01-12)
7. fix `APPEND` and `PREPEND` functions (v0.4.3.3b 2025-01-12)
8. fix `INCR` and `INCRBY` edge cases, incl rollover/under (v0.4.3.3b 2025-01-13)

+++ src/keys.h
@@ -39,10 +39,10 @@ static art::node_ptr leaf_numeric_update(UT &l, const art::node_ptr &old, UT by)
 
         auto old = l;
         l += by;
-        if (by > 0 && l < old) {
+        if ((long long)by > 0ll && l < old) {
             return nullptr;
         }
-        if (by < 0 && l > old) {
+        if ((long long)by < 0ll && l > old) {
             return nullptr;
         }
         auto s = std::to_string(l);
